cleanup(angular): rearrange validation helper function
