+++ packages/angular/src/generators/scam-pipe/lib/validate-options.ts
@@ -1,9 +1,11 @@
 import type { Tree } from '@nx/devkit';
-import { checkPathUnderProjectRoot } from '../../utils/path';
-import { validateProject } from '../../utils/validations';
+import {
+  validatePathIsUnderProjectRoot,
+  validateProject,
+} from '../../utils/validations';
 import type { Schema } from '../schema';
 
 export function validateOptions(tree: Tree, options: Schema): void {
   validateProject(tree, options.project);
-  checkPathUnderProjectRoot(tree, options.project, options.path);
+  validatePathIsUnderProjectRoot(tree, options.project, options.path);
 }
