+++ packages/angular/src/generators/utils/validations.ts
@@ -1,6 +1,12 @@
 import type { Tree } from '@nx/devkit';
-import { getProjects, stripIndents } from '@nx/devkit';
+import {
+  getProjects,
+  normalizePath,
+  readProjectConfiguration,
+  stripIndents,
+} from '@nx/devkit';
 import { lt } from 'semver';
+import { pathStartsWith } from './path';
 import { getInstalledAngularVersionInfo } from './version-utils';
 
 export function validateProject(tree: Tree, projectName: string): void {
@@ -28,5 +34,29 @@ export function validateStandaloneOption(
   if (lt(installedAngularVersion, '14.1.0')) {
     throw new Error(stripIndents`The "standalone" option is only supported in Angular >= 14.1.0. You are currently using "${installedAngularVersion}".
     You can resolve this error by removing the "standalone" option or by migrating to Angular 14.1.0.`);
+  }
+}
+
+export function validatePathIsUnderProjectRoot(
+  tree: Tree,
+  projectName: string,
+  path: string
+): void {
+  if (!path) {
+    return;
+  }
+
+  const { root } = readProjectConfiguration(tree, projectName);
+
+  let pathToComponent = normalizePath(path);
+  pathToComponent = pathToComponent.startsWith('/')
+    ? pathToComponent.slice(1)
+    : pathToComponent;
+
+  if (!pathStartsWith(pathToComponent, root)) {
+    throw new Error(
+      `The path provided (${path}) does not exist under the project root (${root}). ` +
+        `Please make sure to provide a path that exists under the project root.`
+    );
   }
 }
